Update dependencies and enable easily running the script from a live disc
Through the addition of 2 new variables it is now really easy to update the script
so that it can be run via a Linux live disc and still backup all the data.
Also there are now instructions on what to input in the terminal to run the script.

Also fix bugs in the pacman plugin and copy files plugin.